+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v41.0.0-SNAPSHOT - unreleased
 
+### 🎁 New Features
+
+* `Column` class now supports a `sortValue` property. SortValue can be a string to represent another
+  field to sort by for the column, or a function that produces a value for each cell to use when sorting.
+   The values produced by this property will also be passed to the custom comparator if it is defined
+
 ### ⚙️ Technical
 * Improvements to exception serialization to better handle `LocalDate` and similar custom JS classes.