+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 🎁 New Features
 
+* `ErrorMessage` will read its 'error' directly from its context model, if not provided by the
+dedicated `error` prop.
+* `ExpandCollapseButton` now works for grouped grids in addition to tree grids.
 * `Column` class now supports a `sortValue` property. SortValue can be a string to represent another
   field to sort by for the column, or a function that produces a value for each cell to use for sorting.
   The values produced by this property will also be passed to the custom comparator if it is defined.