+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 * `ErrorMessage` will read its 'error' directly from its context model, if not provided by the
 dedicated `error` prop.
 * `ExpandCollapseButton` now works for grouped grids in addition to tree grids.
-
+* `Column` class now supports a `sortValue` property. SortValue can be a string to represent another
+  field to sort by for the column, or a function that produces a value for each cell to use for sorting.
+  The values produced by this property will also be passed to the custom comparator if it is defined.
 
 ### ⚙️ Technical
 * Improvements to exception serialization to better handle `LocalDate` and similar custom JS classes.